Split CollectionInfoCard from Collection page
