+++ src/pages/admin/Collection/Collection.tsx
@@ -1,10 +1,8 @@
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
-import { Tabs, TabsList } from "@/components/ui/tabs";
-import { TabsTrigger } from "@radix-ui/react-tabs";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { useEffect, useState } from "react";
-import { BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, ResponsiveContainer } from "recharts";
 import { useGetMyCollections } from "@/hooks/collection";
 import { CollectionData } from "@/types/collection";
+import CollectionInfoCard from "./CollectionInfoCard";
 
 export default function Collection() {
   const [currentCollection, setCurrentCollection] = useState<CollectionData>();
@@ -15,20 +13,10 @@ export default function Collection() {
   });
 
   useEffect(() => {
-    // collections.forEach((c) => console.log("collection", c));
     if (collections) {
       setCurrentCollection(collections[0]);
     }
   }, [collections, isPending]);
-
-  // Sample data for the chart - replace with actual data
-  const chartData = [
-    { name: "Layer 1", value: 400 },
-    { name: "Layer 2", value: 300 },
-    { name: "Layer 3", value: 200 },
-    { name: "Layer 4", value: 278 },
-    { name: "Layer 5", value: 189 },
-  ];
 
   return (
     <div className="flex flex-col gap-4 p-4">
@@ -75,139 +63,7 @@ export default function Collection() {
           </div>
         </div>
       </section>
-
-      <section className="grid grid-cols-1 gap-4 md:grid-cols-2">
-        {/* Left Column - Collection Image and Information */}
-
-        {/* Collection Card */}
-        {currentCollection && (
-          <div className="flex flex-col gap-4">
-            {/* Collection Image */}
-            <Card>
-              <CardHeader>
-                <CardTitle>Collection Information</CardTitle>
-              </CardHeader>
-              <CardContent>
-                <div className="aspect-video w-full overflow-hidden rounded-md">
-                  <img className="h-full w-full object-cover" src="/DOKPAMI.png" />
-                </div>
-                <div className="mt-4 space-y-2 overflow-auto">
-                  <h1 className="text-xl font-bold">
-                    {currentCollection.objectData.content.fields.base_type.fields.type}
-                  </h1>
-                  <p className="font-semibold">Collection Configs</p>
-                  {currentCollection.dynamicFieldData.map((data) => {
-                    if (!("name" in data.content.fields.value.fields)) return;
-                    return (
-                      <div>
-                        <p>
-                          {data.content.fields.value.fields.name}:{" "}
-                          {data.content.fields.value.fields.content}
-                        </p>
-                      </div>
-                    );
-                  })}
-                </div>
-              </CardContent>
-            </Card>
-
-            {/* Collection Information */}
-            <Card>
-              <CardHeader>
-                <CardTitle>Details</CardTitle>
-              </CardHeader>
-              <CardContent>
-                <div className="space-y-4">
-                  <div className="overflow-auto">
-                    <h3 className="font-medium">Property Types</h3>
-                    {currentCollection.objectData.content.fields.property_types.fields.contents.map(
-                      (data) => (
-                        <p className="text-muted-foreground text-sm">
-                          {JSON.stringify(data.fields.type)}
-                        </p>
-                      )
-                    )}
-                  </div>
-                  <div>
-                    <h3 className="font-medium">Layer Types</h3>
-                    {currentCollection.objectData.content.fields.layer_types.fields.contents.map(
-                      (data) => (
-                        <p className="text-muted-foreground text-sm">
-                          {JSON.stringify(data.fields.type)}
-                        </p>
-                      )
-                    )}
-                  </div>
-                  <div>
-                    <h3 className="font-medium">Ticket Types</h3>
-                    {currentCollection.objectData.content.fields.ticket_types.fields.contents.map(
-                      (data) => (
-                        <p className="text-muted-foreground text-sm">
-                          {JSON.stringify(data.fields.type)}
-                        </p>
-                      )
-                    )}
-                  </div>
-                </div>
-              </CardContent>
-            </Card>
-          </div>
-        )}
-
-        {/* Right Column - Layer Items and Chart */}
-        {currentCollection && (
-          <div className="flex flex-col gap-4">
-            {/* Layer Items */}
-            <Card className="flex-1">
-              <CardHeader>
-                <CardTitle>Layer Items</CardTitle>
-                <CardDescription>Items by layer</CardDescription>
-              </CardHeader>
-              <CardContent>
-                {/* <Tabs defaultValue={"layers"} className="w-full" onValueChange={setActiveTab}> */}
-                <Tabs defaultValue={"layers"} className="w-full">
-                  <TabsList className="w-full">
-                    {/* <TabsTrigger value="no-layers">No Layers</TabsTrigger> */}
-                    {currentCollection.objectData.content.fields.layer_types.fields.contents
-                      .length > 0 ? (
-                      currentCollection.objectData.content.fields.layer_types.fields.contents.map(
-                        (l) => (
-                          <TabsTrigger key={l.fields.type} value={l.fields.type}>
-                            {l.fields.type}
-                          </TabsTrigger>
-                        )
-                      )
-                    ) : (
-                      <TabsTrigger value="no-layers">No Layers</TabsTrigger>
-                    )}
-                  </TabsList>
-                </Tabs>
-              </CardContent>
-            </Card>
-
-            {/* Chart */}
-            <Card>
-              <CardHeader>
-                <CardTitle>Distribution Chart</CardTitle>
-                <CardDescription>Item distribution across layers</CardDescription>
-              </CardHeader>
-              <CardContent>
-                <div className="h-[300px] w-full">
-                  <ResponsiveContainer width="100%" height="100%">
-                    <BarChart data={chartData}>
-                      <CartesianGrid strokeDasharray="3 3" />
-                      <XAxis dataKey="name" />
-                      <YAxis />
-                      <Tooltip />
-                      <Bar dataKey="value" fill="#8884d8" />
-                    </BarChart>
-                  </ResponsiveContainer>
-                </div>
-              </CardContent>
-            </Card>
-          </div>
-        )}
-      </section>
+      {currentCollection && <CollectionInfoCard collection={currentCollection} />}
     </div>
   );
 }
